% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{data_clima}
\alias{data_clima}
\title{Climatic data}
\usage{
data_clima(data_sig, ifn, db = ifn_connect(), ...)
}
\arguments{
\item{data_sig}{tbl connection, generally the result of data_sig function}

\item{ifn}{character indicating the IFN version to load}

\item{db}{database connection}

\item{...}{filter arguments to the dplyr::filter step}
}
\description{
Function to load the climatic data, from the db tables
}
\details{
This function loads the climatic data for those plots selected in data_sig
}
