% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{data_sig}
\alias{data_sig}
\title{SIG data getter}
\usage{
data_sig(ifn, db = ifn_connect(), ...)
}
\arguments{
\item{ifn}{character indicating the IFN version to load}

\item{db}{database connection}

\item{...}{filter arguments to the dplyr::filter step}
}
\description{
Function to load the SIG data
}
\details{
This function does NOT collect the data from the db, only performs the steps
necessary to retrieve and filter the data if necessary from the db. The
collect step must be performed with the result of the function
}
