% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{ifn_connect}
\alias{ifn_connect}
\title{Creating a connection to the IFN SQL database}
\usage{
ifn_connect(user = "guest", password = "guest", dbname = "ifndb",
  idleTimeout = 3600000)
}
\arguments{
\item{user}{character}

\item{password}{character}

\item{dbname}{character}

\item{idleTimeout}{numeric}
}
\description{
Wrapper of \code{\link{dbPool}} to connect to the IFN database
}
\details{
By default (no arguments supplied) it allows to access as a guest to the
database
}
