% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{summarise_polygons}
\alias{summarise_polygons}
\title{Summarise the polygons}
\usage{
summarise_polygons(data_core, polygon_group, func_group = "",
  .funs = dplyr::funs(mean(., na.rm = TRUE), stats::sd(., na.rm = TRUE),
  min(., na.rm = TRUE), max(., na.rm = TRUE), stats::median(., na.rm = TRUE),
  q95 = stats::quantile(., probs = 0.95, na.rm = TRUE), dplyr::n()), ...)
}
\arguments{
\item{data_core}{tbl con, usually obtained from data_core() function}

\item{polygon_group}{character indicating the polygon group to summarise}

\item{func_group}{character indicating the functional group to also summarise,
if any.}

\item{.funs}{funs (dplyr) object with the functions to summarise}

\item{...}{filter arguments}
}
\description{
Function to summarise variable values for each selected polygon
}
\details{
This function takes the core data for the plots and calculate the summaries
based on the polygon kind selected. It also can be filtered before summarising
in the case of dominant functional groups
}
