% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{ifn_connect}
\alias{ifn_connect}
\title{Creating a connection to the IFN SQL database}
\usage{
ifn_connect()
}
\description{
Wrapper of \code{\link{dbPool}} to connect to the IFN database
}
\details{
There is no arguments, this is used to access as a guest to the database
}
